apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT[crypto]==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
email-validator==2.2.0
httpx==0.28.1
beautifulsoup4==4.12.3
redis==5.2.0
